fixed indent and added visualization
